+++ frontend/src/features/map/useVisualizeMarker.ts
@@ -1,3 +1,48 @@
-const useVisualizeMarker = () => {};
+import { useEffect } from "react";
+
+interface LatLng {
+  lat: number;
+  lng: number;
+}
+interface UseVisualizeMarkerProps {
+  map: any;
+  path: LatLng[];
+}
+const useVisualizeMarker = ({ map, path }: UseVisualizeMarkerProps) => {
+  useEffect(() => {
+    if (!map) return;
+
+    const getMarkerType = (
+      index: number,
+      length: number,
+    ): keyof typeof imageSrc => {
+      if (index === 0) return "start";
+      if (index === length - 1) return "end";
+      return "middle";
+    };
+
+    const imageSrc = {
+      start: "/src/assets/icons/marker-start.svg",
+      middle: "/src/assets/icons/marker-default.svg",
+      end: "/src/assets/icons/marker-end.svg",
+    };
+
+    for (let i = 0; i < path.length; i++) {
+      const markerType = getMarkerType(i, path.length);
+      const imageSize = new window.kakao.maps.Size(32, 32);
+      const markerImage = new window.kakao.maps.MarkerImage(
+        imageSrc[markerType],
+        imageSize,
+      );
+
+      new window.kakao.maps.Marker({
+        map,
+        position: new window.kakao.maps.LatLng(path[i].lng, path[i].lat),
+        title: "User Marker",
+        image: markerImage,
+      });
+    }
+  }, [map]);
+};
 
 export default useVisualizeMarker;
